requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
drf-yasg==1.20.0
firebase-admin==2.13.0
firebase-orm==0.1.29
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.4
phonenumbers==8.12.13
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
